requirements.txt
python-dotenv==1.1.0
requests==2.32.3
fastapi==0.115.12
uvicorn==0.34.2
mangum==0.19.0
firebase-admin==6.8.0
python-multipart==0.0.20
authlib==1.5.2
passlib[bcrypt]==1.7.4
python-jose==3.4.0
cryptography==44.0.3
httpx==0.28.1
sentence-transformers==4.1.0
numpy==2.2.5